+++ .github/actions/tests.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+set -e
+
+npx cypress run  --headless --browser chrome  --config '{"specPattern":["plugins/blocks/languageToggleByFlag/cypress/tests/functional/*.cy.js"]}'